Revoking refresh token moved to AuthService. Updated user service injections and usages
