+++ Product.Application/ServiceInterfaces/IUserService.cs
@@ -1,6 +1,4 @@
-﻿using Product.Application.Dto;
-using Product.Domain.Dto;
-using Product.Domain.Entity;
+﻿using Product.Domain.Dto;
 using Product.Domain.Result;
 
 namespace Product.Application.ServiceInterfaces;
@@ -10,5 +8,4 @@ public interface IUserService
 	Task<Response<UserDtoToFrontEnd>> GetUserAsync(int userId, CancellationToken cancellationToken);
 	Task<Response<int>> RemoveUserAsync(int userId, CancellationToken cancellationToken);
 	Task<Response<UserDtoToFrontEnd>> UpdateUserAsync(UserToUpdateDto userUpdateData, int userId, CancellationToken cancellationToken);
-	Task<Response<TokenDto>> Refresh(RefreshTokenRequestDto refreshDto, CancellationToken cancellationToken);
 }
