+++ Product.Infrastructure/Implementations/AuthService.cs
@@ -20,7 +20,7 @@ public class AuthService : IAuthService
     private readonly IAdministratorRepository _adminRepository;
 
     public AuthService(IUserRepository userRepository, IPasswordHasher passwordHasher, IJwtTokenService jwtTokenService,
-        IRefreshTokenRepository refreshTokenRepository, IVendorUserRepository vendorUserRepository, 
+        IRefreshTokenRepository refreshTokenRepository, IVendorUserRepository vendorUserRepository,
         IOperatorUserRepository operatorUserRepository, IAdministratorRepository administratorRepository)
     {
         _userRepository = userRepository;
@@ -31,8 +31,8 @@ public class AuthService : IAuthService
         _vendorUserRepository = vendorUserRepository;
         _operatorUserRepository = operatorUserRepository;
     }
-    
-    public async Task<Response<TokenDto>> LoginAdministrator(UserLoginDto userData, 
+
+    public async Task<Response<TokenDto>> LoginAdministrator(UserLoginDto userData,
         CancellationToken cancellationToken = default)
     {
         var admin = await _adminRepository.GetByEmailAsync(userData.Email, cancellationToken);
@@ -66,6 +66,43 @@ public class AuthService : IAuthService
         };
     }
 
+    public async Task<Response<TokenDto>> Refresh(RefreshTokenRequestDto refreshDto,
+        CancellationToken cancellationToken = default)
+    {
+        var existingToken = await _refreshTokenRepository.GetByTokenAsync(refreshDto.RefreshToken, cancellationToken);
+
+        if (existingToken == null || !_jwtTokenService.Validate(existingToken))
+            return new Response<TokenDto>
+            {
+                ErrorMessage = "Refresh token expired",
+                ErrorCode = (int)ErrorCodes.InvalidRefreshToken
+            };
+
+        await _refreshTokenRepository.RevokeAsync(existingToken, cancellationToken);
+
+        var userId = existingToken.UserId;
+
+        var user = await _userRepository.GetByIdAsync(userId, cancellationToken);
+
+        var userClaims = user.MapUserToClaimDto();
+
+        var newToken = _jwtTokenService.GenerateToken(userClaims);
+
+        var refreshToken = new RefreshToken
+        {
+            Token = newToken.RefreshToken,
+            UserId = userId,
+            ExpiresAt = DateTime.UtcNow.AddDays(7)
+        };
+
+        await _refreshTokenRepository.CreateAsync(refreshToken, cancellationToken);
+
+        return new Response<TokenDto>
+        {
+            Data = newToken
+        };
+    }
+
     public async Task<Response<TokenDto>> Login(UserLoginDto userData, CancellationToken cancellationToken = default)
     {
         var user = await _userRepository.GetByEmailAsync(userData.Email, cancellationToken);
@@ -91,27 +128,27 @@ public class AuthService : IAuthService
         }
 
         var userClaims = user.MapUserToClaimDto();
-		
+
         var token = _jwtTokenService.GenerateToken(userClaims);
-		
+
         var refreshToken = new RefreshToken
         {
             Token = _jwtTokenService.GenerateRefreshToken(),
             UserId = user.Id,
             ExpiresAt = DateTime.UtcNow.AddDays(7)
         };
-		
+
         await _refreshTokenRepository.CreateAsync(refreshToken, cancellationToken);
-		
+
         token.RefreshToken = refreshToken.Token;
-		
+
         return new Response<TokenDto>
         {
             Data = token,
         };
     }
 
-    public async Task<Response<UserDtoToFrontEnd>> RegisterUser(UserRegistrationDto registrationData, 
+    public async Task<Response<UserDtoToFrontEnd>> RegisterUser(UserRegistrationDto registrationData,
         CancellationToken cancellationToken = default)
     {
         var userByEmail = await _userRepository.GetByEmailAsync(registrationData.Email, cancellationToken);
@@ -123,12 +160,12 @@ public class AuthService : IAuthService
                 ErrorCode = (int)ErrorCodes.UserWithThisEmailAlreadyExists
             };
         }
-		
+
         if (!registrationData.IsOperator)
         {
             var newVendor = registrationData.MapVendorUserFromDto();
             newVendor.PasswordHash = _passwordHasher.HashThePassword(registrationData.Password);
-            
+
             await _vendorUserRepository.CreateAsync(newVendor, cancellationToken);
             var vendorUserDto = newVendor.MapToFrontEndDto();
             return new Response<UserDtoToFrontEnd>
@@ -136,12 +173,12 @@ public class AuthService : IAuthService
                 Data = vendorUserDto
             };
         }
-		
+
         var newOperator = registrationData.MapOperatorUserFromDto();
         newOperator.PasswordHash = _passwordHasher.HashThePassword(registrationData.Password);
-        
+
         await _operatorUserRepository.CreateAsync(newOperator, cancellationToken);
-		
+
         var operatorUserDto = newOperator.MapToFrontEndDto();
 
         return new Response<UserDtoToFrontEnd>
