+++ Product.WebApi/Controllers/AccountController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Product.Application.Dto;
-using Product.Application.Interfaces;
 using Product.Application.ServiceInterfaces;
 using Product.Domain.Dto;
 using Product.Domain.Result;
@@ -118,7 +117,7 @@ public class AccountController : ControllerBase
 	public async Task<ActionResult<Response<TokenDto>>> RefreshToken([FromBody] RefreshTokenRequestDto tokenRequestDto,
 		CancellationToken cancellationToken)
 	{
-		var response = await _userService.Refresh(tokenRequestDto, cancellationToken);
+		var response = await _authService.Refresh(tokenRequestDto, cancellationToken);
 		if (response.IsSuccess)
 		{
 			return Ok(response);
